requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Django==3.1.7
django-dotenv==1.4.2
django-money==1.3.1
djangorestframework==3.12.4
download-gitignore==2.0.1
idna==3.1
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
py-moneyed==1.2
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
